+++ src/sub-clients/contentPartnerContracts/model/ContentPartnerContractRoyaltySplit.ts
@@ -1,4 +1,0 @@
-export interface ContentPartnerContractRoyaltySplit {
-  streaming: number;
-  download: number;
-}