+++ src/sub-clients/contentPartnerContracts/resources/ContentPartnerContractDatesResource.ts
@@ -1,4 +1,0 @@
-export interface ContentPartnerContractDatesResource {
-  start?: string;
-  end?: string;
-}