+++ src/sub-clients/contentPartnerContracts/resources/ContentPartnerContractRoyaltySplitResource.ts
@@ -1,4 +1,0 @@
-export interface ContentPartnerContractRoyaltySplitResource {
-  streaming: number;
-  download: number;
-}