+++ src/sub-clients/contracts/model/ContractRoyaltySplit.ts
@@ -1,0 +1,4 @@
+export interface ContractRoyaltySplit {
+  streaming: number;
+  download: number;
+}