+++ src/sub-clients/contracts/resources/ContractCostsResource.ts
@@ -1,4 +1,4 @@
-export interface ContentPartnerContractCostsResource {
+export interface ContractCostsResource {
   minimumGuarantee: number[];
   upfrontLicense: number;
   technicalFee: number;