+++ src/sub-clients/contracts/resources/ContractDatesResource.ts
@@ -1,0 +1,4 @@
+export interface ContractDatesResource {
+  start?: string;
+  end?: string;
+}