+++ src/sub-clients/contracts/resources/ContractRoyaltySplitResource.ts
@@ -1,0 +1,4 @@
+export interface ContractRoyaltySplitResource {
+  streaming: number;
+  download: number;
+}